Fix/ub-1432 : check correctly pv existance during uninstall (#246)
during ubiquity in install process we are waiting for all objects to be deleted.
one of them is the PV object.
the problem is that there are currently some cases which cause the PV to remain in kuberentes even though the PVC is deleted (idempotent delete issue)
because of this it is needed to find another way to get the PV name instead of getting it from the PVC as we do now. so the solution here was to use the name in the ubiquity-configmap to check if a PV with this name remains.
if the PV is still there we will stop the un-install process on this step no matter how many times it was invoked by the user until someone will intervene and remove the PV and fix the issue.