Create credential secret and storage class only for scale or for scbe
for scale the secret definision is in spectrumscale-credentials-secret.yml and the first storage class definision is storage-class-spectrumscale.yml
but for scbe its different objects as follow:
for scbe the secret definision is in scbe-credentials-secret.yml and the first storage class definision is storage-class.yml
So in order to support it in the helm chart we should create them only for scale or for scbe based on the values details. Means if fqdn of scale define then we go for scale, if its for scbe then we go for scbe objects.